Use local variables for daemon.json template
Instead of using bare lists and dicts this change converts variables
into json for use with daemon.json. jinja2 does have an ability to do
this in side of templates, however it is only available in 2.9+. Until
then the use of ansible's to_json in the playbook is being used.

Refs:
- http://jinja.pocoo.org/docs/2.9/templates/#tojson
- http://docs.ansible.com/ansible/playbooks_filters.html#filters-for-formatting-data