Don't set-up origin repositories if they've already been configured
CentOS ships the 'centos-release-openshift-origin' package which is
used to set up the repository and the key. If this package has been
installed prior to running openshift-ansible (i.e, to install
ansible or openshift-ansible from RPM), it leads to the repository
being set up twice in different files.

We're now matching the name of the files and since they are identical,
it will stay idempotent while not configuring the repositories twice.