Scale logging deployments while paused for rollout sanity
When we currently create the set of logging `DeploymentConfig`s, we
create them with zero desired replicas. This causes the deployment to
immediately succeed as there is no work to be done. This inhibits our
ability to use nice CLI UX features like `oc rollout status` to monitor
the logging stack deployments. If we want to keep the current separation
of `DeploymentConfig` creation and scaling, we can create the configs in
the paused state and unpause them as a condition of cluster start.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>