RPM workaround for the move of cert playbooks
The playbooks/certificate_expiry directory is being moved inside the
openshift-checks tree, and a symlink there is put in place for backwards
compatibility. RPM does not handle this well, so this expands the section
of the specfile that already contains a workaround for this situation to
also handle this move.