oc_atomic_container: Workaround for invalid json from atomic command
When no other containers are present and one attempts to list containers
via the atomic command with the --json flag an empty string is returned.
If a json.loads is used on this value an error is raised. This change
adds a workaround to fall back to '[]' when an empty string is returned
from the atomic containers command.