memory check: use GiB/MiB and adjust memtotal
fixes https://bugzilla.redhat.com/show_bug.cgi?id=1455884

Various things reserve memory such that memtotal is quite lower than the
actual physical RAM of the system. It's larger as RAM increases but it's
not really proportional so I just added a flat 1GiB adjustment in the
comparison. This ought to "pass when it's close enough."