Changed the Openstack playbooks to use the native os_stack module
This changes the openstack playbooks to use the native ansible os_stack
module for creating and managing the heat stack for Openshift, rather than
manually invoking the heat command line client. This is because

* The heat command line client itself is depreciated (in favor of the
  unified python-openstackclient

* The output of the command seems to have changed which breaks the
  current method of scraping the output to obtain the output values

* We get some greater flexibility in using the ansible modules as
  they natively hook onto os-cloud-config for cloud credentials

This means that you need to have python-shade installed (as the ansible
openstack modules depend on it) along with the normal Openstack clients.