Rename container image to origin-ansible / ose-ansible
As discussed in #4279 the container image is being renamed to
openshift/origin-ansible (upstream) and openshift3/ose-ansible (downstream) for
the "oc cluster up" use case, and more generally to make it fit the
imageConfig.format configuration option.

This updates references to the image name accordingly.

Also updating references to playbook2image to account for recent moves of the
repo/image.