Suggest installing java-1.8.0-openjdk-headless and httpd-tools (#442)
* Suggest installing java-1.8.0-openjdk-headless

java-1.8.0-openjdk-headless is required when installing metrics.  Failure to do so results in a confusing message:

fatal: [osacpr00-master-p8tg]: FAILED! => {"changed": false, "failed": true, "msg": "'keytool' is unavailable. Please install java-1.8.0-openjdk-headless on the control node"}

The first few times I read this I misunderstood that this was a local check for the machine that was trying to bring up the cluster.  The message admittedly says "control node" right there on the tin but I guess I didn't make the connection.  Regardless, seems better to add it to the list of required packages.

* Create README.md