Fix possible access to undefined variable
g_glusterfs_hosts is verified to be defined in the beginning of the
playbook.
g_glusterfs_registry_hosts is only defined when
playbooks/byo/openshift-cluster/cluster_hosts.yml is loaded, which seems
to not always be the case.

Instead of trying to find all entrypoints where cluster_hosts should be
loaded to defined the variable, simply use the `default()` filter to
consider it empty when not defined.