docker_storage check: make vgs return sane output
fix bug 1464974 https://bugzilla.redhat.com/show_bug.cgi?id=1464974

Specify --units on vgs call. In my testing with lvm 2.0.2.171(2) on
RHEL Atomic Host 7.4, this turned a response of "<4.07g" into "4.07g"
which should resolve the issue. I haven't found what the "<" is for
in the first place but I'm thinking this should at least be a safe
change.