Only store failures that were not ignored.
In the past, health checks were implemented with ignore_errors: True in
the playbook level, requiring us to store all failures, ignored or not,
so that we could report on all failed checks.

Now checks are run from a single action plugin entry point, without
ignoring errors (all errors are aggregated via the action plugin).

Since the integration of the openshift_health_checker role with the
install playbook, failure summaries are part of the output of a lot more
calls to ansible-playbook. We shall report only failures that caused the
execution to stop, as ignored failures in the summary only serve to
confuse users.