openshift_checks: fix execute_module params
Fix where execute_module was being passed task_vars in place of tmp
param. Most modules don't seem to use either and so this doesn't fail;
but under some conditions (perhaps different per version of ansible?) it
tried to treat the dict as a string and came back with a python stack
trace.