Set ansible_become for the OSEv3 group
Because openshift-ansible requires root on the cluster nodes, but it
doesn't explicitly set it in the playbooks (like we do), let's set it
in our inventory instead of requiring to pass `--become` to
`ansible-playbook`.

That will simplify the installation steps as well as let us include
the provisioning and openshift-ansible playbooks in a single playbook.