Misc cleanup and input normalization
Clean up openshift-checks playbooks

- The referenced issue was closed, seems like a wontfix. Just use the
old syntax and make no noise about it.
- Remove inconsistent use of tags: if tags are passed in the command
line, there is no point in evaluating groups if the checks won't be run
anyway.
- Remove temporal comment, easy to get out of sync.

Make LoggingCheck.run return the correct type

The run method is expected to return a dict. Even though we should not
run LoggingCheck by itself, it is still possible to do it and without
this change we get an unhandled exception.

Clean up unnecessary quotes

Normalize list of checks passed to action plugin

This allows users to pass either a string with comma-separated check
names or a proper list of strings, whatever is more convenient in the
context.

For consistency, the same normalization is applied to the
'openshift_disable_check' variable.