package_version check: tolerate release version 3.7
Addresses issue https://github.com/openshift/openshift-ansible/issues/4967

For now, any version >= 3.6 is handled as if it were 3.6. We may want to
keep that or fine-tune it later.

Also, the ovs_version check is not updated. This is a post-install
health check (does not block install/upgrade) with an update already in
progress so will be addressed there.