Origin image build: add oc client
The oc client was left out with the move away from playbook2image.
However it is needed in the image for the `oc cluster up` use case for
installing logging or metrics.

This change adds oc via the latest signed and released client RPM,
which requires adding the centos-openshift-origin repo to the image.