Flatten docker into one role
This commit is intended to flatten the various docker roles
into a single docker role and remove all meta-depends of the
various docker roles in all other roles.

This commit also modifies other roles to not have meta-depends
on the docker role.  This requires that docker role be run
specifically on all host groups that require docker.  The
docker role is now called directly during setup of masters,
nodes, and etcd hosts.