Fix registry auth task ordering
Currently, registry authentication credentials are not
produced until after docker systemd service files are
created.

This commit ensures the credentials are
created before the systemd service files to ensure
the proper boolean is set to include the read-only
mount of credentials inside containerized nodes and
masters.

This commit also ensures docker authentication credentials
are mounted into master containers in case of a single
master deployment.  This functionality was missing previously.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1316341