Empty ssh (#729)
* Make `openstack_private_ssh_key` optional

Before this, the deployer could not reasonably rely on their own SSH
configuration or e.g. using the `--private-key` option to
ansible-playbook because we always wrote the `ansible_private_key_file`
value in the static inventory.

This change makes the `openstack_private_ssh_key` variable truly
optional: if it's not set, the static inventory will not configure the
SSH key and will just rely on the existing configuration.

* Update the openstack e2e CI

It no longer sets the SSH keys explicitly -- which should just work with
the previous commit.

* Put back the `openstack_ssh_public_key` in CI

This is the option we actually need to keep. This sholud fix the CI
failures.