Support for scaling nodes with prebuilt AMI
This commit provides the ability to bring your own,
prebuilt AMI to scale out nodes.  This requires
the nodes be created either by hand, or using
the play in playbooks/aws/openshift-node/provision_nodes.yml

Once the nodes are provisioned, they must be added to the
'new_nodes' group in the inventory.

This commit also provides refactoring, bug fixes,
and improvements for ami_build and related code.