openshift_checks: Fix incorrect list cast
docker_image_availability casted openshift_docker_additional_registries
to a list using the list() function. If a string was returned (IE: only
a single registry added) the result would be the string split up by
component characters. This change forces a string result from get_var to
be placed inside a list. If the result is anything BUT a string the
original list() function is called on the result.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1497274

Signed-off-by: Steve Milner <smilner@redhat.com>