Simplify dns management
Switches dnsmasq to listen on all interfaces other than lo by default.
If you specify a list of interfaces to listen on it will use those
instead.

In 3.7 installs it relies on the node to send dbus messages eliminating
additional configuration files.

Some minor fixes to the dispatcher script to simplifying things.