Streamline the OpenStack provider README
This moves all the extra configuration options and deployment notes to a
new `advanced-configuration.md` file and keeps the README much shorter.

The README now presents the simplest workflow with minimal configuration
and manual steps on part of the deployer.

The advanced configuration is in need of a little more cleanup, but we
can do that in another pull request.