Use "requests" for CPU resources instead of limits
We now use a CPU request to ensure logging infrastructure pods are
not capped by default for CPU usage. It is still important to ensure
we have a minimum amount of CPU.

We keep the use of the variables *_cpu_limit so that the existing
behavior is maintained.

Note that we don't want to cap an infra pod's CPU usage by default,
since we want to be able to use the necessary resources to complete
it's tasks.

Bug 1501960 (https://bugzilla.redhat.com/show_bug.cgi?id=1501960)