Fix missing docker option signature-verification
Currently, docker 1.12 can be configured to use the
run-time parameter of 'signature-verification'

By default, rpm installation of docker results in
'--signature-verification=false' being added to
OPTIONS in /etc/sysconfig/docker

Currently, openshift-ansible does not preserve that
value.  This can cause docker to be unable to pull
images from a previously working source due to a
change in configuration.

This commit adds the option with the default
as provided by the rpm installation.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1502560