reconcile registry-console and docker_image_availability
Fixes bug 1497310
https://bugzilla.redhat.com/show_bug.cgi?id=1497310

The registry console is a special case in more than one way. This adds
logic to incorporate the openshift_cockpit_deployer_* variables into
determining what its image will be in docker_image_availability.

Along the way I noticed the origin and enterprise templates for this
were not consistent. Now they are, and the example hosts file is updated.