Fix preupgrade authorization objects are in sync minor versions
Currently, we check that upgrade target is less than 3.8,
but this will break for minor upgrades.

This commit set's a fact early in the upgrade process to
deterime what the currently installed version on the
first master is.

This fact is used to determine if our currently installed
version is less than 3.7.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1508301