Run registry auth after docker restart
Currently, docker login may fail if a proxy is added to the config
but docker is already running.

This is due to the fact that 'docker login' must have a functioning
docker.service running (with valid network connection) to complete.

Currently, handlers restart the docker service at the end of
the role.  This doesn't allow for updating proxy settings before
running docker login.

This commit moves 'docker login' command after flushing handlers.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1511869