papr: auto-detect image tag to use and bump to f27
Current PAPR tests are failing because we're using an outdated image
tag. Switch to auto-magically determining which image tag to use based
on the branch the PR we're running in is targeting, which lets us not
have to worry about the tag version becoming stale again. This is
similar to what the containerized tests from `aos-cd-jobs` are doing
(i.e. use `git describe --abbrev=0` in the target branch to determine
the image tag to use).

Also bump to the brand new f27 release, which allows us to drop the
temporary workaround for the glibc/libsolv bug.