Use IP addresses for OpenStack nodes
Unlike other cloud providers, OpenStack VMs are not able to resolve each
other by their names. If you try to run the playbooks against nodes
without a pre-created /etc/hosts or a DNS that provides the hostname/ip
resolution, it will fail.

By setting the `openshift_hostname` variable to each node's IP address,
we're able to deploy a functional cluster without running a custom DNS.

It is still possible to provide an external server with nsupdate keys
and have it be populated, but that is no longer a hard requirement.