Fix the env lookup fallback in rhel_subscribe
The role is looking for the usernames, passwords, etc. first in the
shell environment and then in the Ansible vars.

When the environment is empty, however, the lookup returns an empty
string not an undefined variable, so the `default` fallback is never
used and the ansible variables end up being ignored.

By adding `True` to the end of the filter, it will work with any falsey
value (including an empty string) not just undefined variables.