Fix inventory symlinks in origin-ansible container.
When attempting to run openshift-ansible in a kubernetes pod, we
attempted to use a configmap to mount in our inventory to run against.
However kubernetes mounts configmaps/secrets volumes through the use of
relative symlinks. The use of "cp -a" in this script then copies a
relative symlink to /tmp which is now dangling, and ansible will fail to
run.

This patch changes to just copy the file regularly resulting in a value
temp file that is not a symlink.