Fix docker_image_availability checks
This commit ensures that oreg_url is properly templated
by ansible before being consumed in the logic.

This commit also adds a method to the base health check
class to detect if self._templar is none, and return
the appropriate templated/untemplated version of the
variable.  This is mostly for unit tests.