Migrate to static:true for include_role
In Ansible 2.2, the `include_role` directive came into existence as
a Tech Preview. It is still a Tech Preview through Ansible 2.4
(and in current devel branch), but with a noteable change. The
default behavior switched from `static: true` to `static: false`
because that functionality moved to the newly introduced
`import_role` directive (in order to stay consistent with `include*`
being dynamic in nature and `import* being static in nature).

The dynamic include is considerably more memory intensive as it will
dynamically create a role import for every host in the inventory
list to be used. (Also worth noting, there is at the time of this
writing an object allocation inefficiency in the dynamic include
that can in certain situations amplify this effect considerably)

This change is meant to mitigate the pressure on memory for the
Ansible control host while maintaining compatibility with Ansible
2.3 and Ansible 2.4. (Eventually this should be migrated to use
`import_role` everywhere, but that directive is not present in
Ansible 2.3)

Signed-off-by: Adam Miller <admiller@redhat.com>