Don't hardcode the network interface in the openshift_logging_mux role
The openshift_logging_mux role hardcodes the 'eth0' interface alias
for determining the IP address to use for incoming external client
connections.  This will cause the playbook to fail with an undefined
variable error on systems where an 'eth0' interface does not exist.

This patch changes the default IP address for external connections
to use the 'ansible_default_ipv4' fact.  It also allows this to be
overridden by a new 'openshift_logging_mux_external_address' variable.