Define ssd storage class for GCE
GCE provides two storage options - hdd and ssd. During the OpenShift
installation, the default storage class is set to hdd, what is fine, but ssd
storage is not defined. This PR creates the ssd storage class during the
installation on GCE provider.

Let me know if this is not the proper way how to introduce an additional
storage class.