Move the OpenStack dynamic inventory from sample
Keeping the dynamic inventory in the `sample-inventory` alongside the
`group_vars` poses a problem: when we update the inventory, our users
won't get the latest version.

The dynamic inventory should be independent of the OpenShift or
OpenStack configuration and the users should be able to either pass it
explicitly or create a symlink if they want.