openshift_checks: repair adhoc list-checks mode
fixes bug https://bugzilla.redhat.com/show_bug.cgi?id=1509157

docker_image_availability.__init__ was loading a variable that of course
was not defined for localhost, which is how this play runs. Fixed that.
Would be nice if there were a more generic way to solve that class of
problem but I can't think of anything for now that isn't worse.

Also noticed that this play gathers facts for localhost. That's not
necessary for what it's doing, so disabled that.