Ensure atomic_proxies are configured with docker
In some setups, system containers may be used
in conjunction with package_docker.

This commit ensures that the atomic
proxies are configured while using package_docker
and containerized components.

This commit also ensures containerized_groups
are based on openshift_is_containerized instead
of 'containerized' variable directly to
account for possible use of atomic hosts which
users may not specify containerized in inventory
directly.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1535323