Change catalog roles install to use aggregation
Recently it was changed for openshift cluster roles to be reconciled via
the aggregate-to* label, so this is now updated. Rather than generating
portions of the file, simply use the auth reconciler and let it handle
any necessary role updates. Also, since it is theoretically possible for
conflicts to occur, retry reconciliation several times.