Merge pull request #6969 from mjudeikis/gluster-dynamic-invetory-bug
Automatic merge from submit-queue.

dynamic inventory bug when group exists but its empty

@jarrpa output from my test for the gluster block

I use dynamic inventory. And the result of this is that ansible see all possible groups available, but some of them are empty [1]. Because group exists, fist condition is evaluated as true and it fails with the array is an empty error...
```
...
'all': [u'54.144.5.175',
         u'52.200.175.54',
         u'54.237.208.252',
         u'52.55.224.72',
         u'52.207.103.246',
         u'34.227.95.191',
         u'34.238.127.160'],
 u'bastion': [u'34.227.95.191'],
 u'bastion.0': [u'34.227.95.191'],
 u'etcd': [u'34.238.127.160'],
 u'glusterfs': [u'52.200.175.54', u'52.207.103.246', u'54.144.5.175'],
 u'glusterfs.0': [u'52.200.175.54'],
 u'glusterfs.1': [u'52.207.103.246'],
 u'glusterfs.2': [u'54.144.5.175'],
 u'glusterfs_registry': [],
 u'infra': [u'52.55.224.72'],
 u'infra.0': [u'52.55.224.72'],
 u'masters': [u'34.238.127.160'],
 u'masters.0': [u'34.238.127.160'],
...
```