Remove RBAC console template
The console is changing to discover if the template service broker is
running by looking at the service classes on the client. This means the
console no longer needs the RBAC template. It also no longer needs to
manually trigger a rollout after the template service broker is
installed or removed.