Simplify double upgrade version logic
Currently, double upgrade process (3.7 -> 3.9)
for control plane attempts to run openshift_version
role twice to set the appropriate values for
upgrading each major version, 3.8 and 3.9.

This commit instructs openshift_version to
only inquire about the proper settings for 3.9,
and hard-sets the appropriate values for 3.8.

This allows a simplification of the
openshift_version role, allowing for easier
debugging.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1542368