deploy k8s job for applying hawkular-metrics schema
This is just a first pass to test whether I have the ansible stuff set
up correctly.

add template for job and support for upgrading

We cannot use oc apply because the job template has immutable fields.

add a comment

fix indentation

Only delete job if it is already installed

There are a couple minor changes based on PR review. First,
ignore_errors is no longer used with the job deletion. Secondly, a new
variable is registered for listing the jobs. The deletion tasks checks
the output of that variable to determine whether or not the job is
already installed.

clean up when conditional

fix quoting error