Cast string to dict in lib_utils_oo_dict_to_keqv_list
Due to bug in ansible not treating host-vars and group-vars
exactly the same when evaluating a string to see if
it should be cast as a dictionary, some users experience
a failure of lib_utils_oo_dict_to_keqv_list due to
the module receiving a string instead of a dictionary.

This commit attempts to create a string from a
dictionary if the dictionary is not provided.

Ansible bug report: https://github.com/ansible/ansible/issues/36544

Fixes: https://github.com/openshift/openshift-ansible/issues/6897
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1542099