Fix package tasks ordering in OpenStack playbooks
The conditional in the `openshift_repos` include in provision.yml was wrong (the
`rh_subscribed` is defined even when the `rhel_subscribe` role is
skipped.

This removes it entirely, because it's called from the main
prerequisites import anyway.

We also moved any package-related tasks from provision to install,
because the OpenStack nodes may need additional configuration before
they're able to install them (e.g. a RHEL subscription or extra repos
that are not covered by our role.

So `provision.yml` now only creates the nodes and makes sure they're up
and ready for the installation. If the user wants to do anything before
we touch the nodes' contents, they can now do so.