etcd migrate: instead of scaleup playbook etcd server should be started back
master doesn't need to be restarted and etcd URLs updated as new etcd
nodes are being added. There is no need to run scaleup playbook, as etcd
nodes are already added to the cluster.

The migrate procedure now does the following:
* checks if the etcd data needs to be migrated
* makes etcd backup
* stops etcd services on all nodes except the first one
* migrates etcd data on first etcd node
* clears data on other etcd nodes
* updates etcd cluster configuration, updating ETCD_INITIAL_CLUSTER and
ETCD_INITIAL_CLUSTER_STATE
* starts the etcd service again one by one on etcd nodes

Now the only copy of migrated data is on the first etcd cluster, which
would replicate it to other nodes.

After migration is done master configs are updated and services
restarted if needed

Signed-off-by: Vadim Rutkovsky <vrutkovs@redhat.com>