Make broker pods run correct versions on upgrade
This ensures that the DeploymentConfig for the ansible service broker is
updated on upgrade (in addition to first install). Since the deployment config
was already using a strategy type of rolling, removing the skip
condition was all that was necessary to launch updated pods.

In the case of the template service broker the daemonset was correctly being
updated with the new image version, but the update strategy was using the
default OnDelete. Changing the strategy to RollingUpdate ensures the
apiserver pod is upgraded as well.