There are times when we want to re-deploy a single component without re-deploying the entire cluster.
This adds a fact, full_install, to the deploy_cluster.yml playbook.

Individual component playbooks, i.e., openshift-logging, check for
the full_install fact and if it's defined, skip calling the
init/main.yml playbook.

If full_deploy is not defined, the component playbook will call the
init/main.yml playbook and configure the facts necessary to deploy
the component.