origin-ansible-container: OpenStack support
the OpenShift Ansible container documentation in the repo states that it
should support the playbooks in this repo. Unfortunately, due to some
extra dependencies in the openstack case, it failed. This patch aims to
make it easy for playbooks that have extra requirements to be able to
build their own tagged origin-ansible containers.

It would be nice if it could be pushed to docker hub as
openshift/origin-ansible:v3.10.0-openstack

Change-Id: Ibb162ec01a432f6b9d061ede76e9e0e35c64f9f5
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>