Refactor openshift_version setting
Currently, we do lots of manipulation of user-provided
version-related variables.  Often times this manipulation
is undesired and incorrect.

This commit refactors openshift_version to not manipulate
so many things.  We will trust that the user has put
correct variables in their inventory to get the desired
outcome.

Implements: https://trello.com/c/sL9bMo3g/644-refactor-version-handling-to-new-format