GlusterFS: enable modprobe in pods that manage bricks
Bricks are created with an LVM thin-pool to allow LV-snapshots to be
taken. This requires certain device-mapper modules to be loaded in the
kernel. In case modules are missing, the lvm-tools (like lvcreate) will
try to load them. If the modules are not loaded in advance, and loading
fails in the pod, then it will not be possible to provision bricks for a
Gluster Volume and volume creation will fail.

See-also: gluster/gluster-kubernetes#457
Signed-off-by: Niels de Vos <ndevos@redhat.com>