add support for a custom sdn role
This adds support for the user to specify a custom sdn role. There are 2 ways this could be done. However in both methods, you use the feature by specifying `openshift_custom_sdn_role` as the name of a role to be imported during the SDN setup phase.

The first method is to simply take the custom role and drop it into the roles/ directory.

The second method is to use the openshift-ansible repo as an exteral dependency of another repo. This is a more complex solution, but can be used in version control.
First you would add the openshift-ansible repo as a submodule or subtree within the parent repo.
Then you would add the following to a parent repo playbook:

    - import_playbook: openshift/playbooks/deploy_cluster.yml

And then add the following to the parent repo `ansible.cfg`:

    [defaults]
    roles_path = roles

This is so that when the openshift-ansible repo goes to look for the custom role, it can find it.