Set OpenStack VM hostname to the entry in Nova
OpenStack appends a domain suffix (`.novalocal` by default) to the
hostnames of the VMs it creates. This clashes with the new control plane
install tasks that look at the hostname rather than the
`openshift_hostname` variable.

This makes the OpenStack playbooks function again with the new control
plane/bootstrap installation.