Set the master cluster hostname under OpenStack
Depending on the deployment (with or without a loadbalancer, with or
without external DNS) the `openshift_master_cluster_hostname` may be
different. And given the OpenStack's hostname suffix behaviour, it is
almost never set to the right value on its own.

This updates the Heat templates and the dynamic inventory to set the
variable to a value that should always work and it can always be
overridden in the group_vars .