make cassandra snapshots configurable
Auto snapshots are now disabled. The Cassandra image exposes a
TAKE_SNAPSHOT environment variable. It can turned on/off using the
openshift_metrics_cassandra_take_snapshot variable which defaults to
False. When true, the snapshot will be taken during the post startup
lifecycle hook, and the snapshot name will be a timestamp.