Hardcode htpasswd auth provider filename
Previously, we allowed users to specify an arbitrary path
for the htpasswd file.  Due to switching master process
to a pod, we need everything for the master to be present
in /etc/origin/master.

This commit removes the ability to specify the path
of the htpasswd file and hardcodes it to
/etc/origin/master/htpasswd

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1565447