catalog: change image version latest
This sets it to the current latest 3.7 release, which probably won't
change again. Even if it does, the catalog code is highly unlikely to
change, so hard coding this to match the enterprise image labeling I
think is sufficient.

Future branches have the image to use available as a parameter, so this
is a 3.7 specific change only.