Move crio.conf format-control to it's template
Previously, the contents of ``l_insecure_crio_registries`` and
``l_additional_crio_registries`` were pre-formatted from their 'docker'
variable equivalents.  They were encoded as as a giant string of
quoted, comma-separated-values at definition time.  Forming the
contents of those variables for use in ``crio.conf``, in a role's
defaults file required lots of quote-escaping and black-magic-voodoo.

ref: https://github.com/openshift/openshift-ansible/blob/release-3.9/roles/container_runtime/defaults/main.yml#L95

Simplify this greatly by simply accepting ``l_docker_insecure_registries`` and
``l_docker_additional_registries`` directly, as-is (list of registry strings).
Handle the ``crio.conf``-specific formatting, directly in the template using
an in-line for-loop.

This puts format-control in the template's hands, which is a much better
strategy from a maintenance perspective.  If a developer needs to account
for some future ``crio.conf`` change, the necessary formatting logic is in
an obvious, easy-to-find place and in simple terms.

Signed-off-by: Chris Evich <cevich@redhat.com>